requirements.txt
torch
torchvision
tqdm
scikit-learn
pyyaml
matplotlib
Pillow
